fix: CanAutoGen logic (#12779)
